requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt
# Or use e.g. pip install -U Flask if on ifi server (you need to be in user env)

# Flask Framework
Flask #==1.0.2
click # ==8.0.4
itsdangerous # ==2.1.2
Jinja2 # ==3.1.1
MarkupSafe # ==2.1.1
Werkzeug # ==2.0.3
flask-cors

# Flask Packages
# Flask-Login==0.4.0
# Flask-Migrate==2.0.2
# Flask-Script==2.0.5
# Flask-SQLAlchemy==2.4.0
# Flask-WTF==0.14.2
# Flask-User==1.0.1.5

# Basic requirements
numpy
shortuuid
argparse
requests
python-dotenv

# Models and database
#SQLAlchemy # ==1.4.25
marshmallow # ==3.15.0

# Azure
azure-storage-blob
azure-data-tables
azure-storage-queue
azure-communication-email
redis

# Color distance
Pillow

# qr
qrcode[pil]

# payments
stripe